% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRISJbot.R
\name{readRISJbot}
\alias{readRISJbot}
\title{A Reader function generator for RISJbot JSONLines data}
\usage{
readRISJbot(mappings)
}
\arguments{
\item{mappings}{An optional named list mapping input JSONLines field names to
metadata fields in the output documents. If mappings are provided, it must
include a mapping for \code{content}, to indicate which part of the input
JSONLines file should be used for the main text of the output document.

If mappings are not provided, \code{content} is collected from the \code{bodytext}
field, all other input fields are carried over to the output document
with their original names, and the \code{datetimestamp} output field is taken as
the first available of the following: \code{datetimestamp} input field,
\code{modtime} input field, \code{firstpubtime} input field, \code{fetchtime} input field,
or (if all else fails) the current date and time, with a warning.}
}
\value{
A \link[tm:Reader]{tm::Reader} function which returns objects inheriting from
\link[tm:PlainTextDocument]{tm::PlainTextDocument}
}
\description{
A Reader function generator for RISJbot JSONLines data
}
\seealso{
\link{RISJbotSource}
}
