% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm.plugin.risjbot-package.R
\docType{package}
\name{tm.plugin.risjbot-package}
\alias{tm.plugin.risjbot}
\alias{tm.plugin.risjbot-package}
\title{tm.plugin.risjbot: Import Articles from RISJbot Using the 'tm' Text Mining Framework}
\description{
Provides a tm Source function to create corpora from articles
  scraped by the RISJbot webcrawler <https://github.com/pmyteh/RISJbot>. It
  reads JSONLines files including both text and metadata.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/pmyteh/tm.plugin.risjbot}
  \item Report bugs at \url{http://github.com/pmyteh/tm.plugin.risjbot/issues}
}

}
\author{
\strong{Maintainer}: Tom Nicholls \email{tom.nicholls@bsg.ox.ac.uk} (0000-0002-6971-8614)

}
\keyword{internal}
